pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = ""
authors = ["Khwanchanok Chaichanayothinwatcahra <khwanchanok.chc@student.mahidol.edu>"]

[tool.poetry.dependencies]
python = ">3.7, <4"
pytest = "^7.2.1"
postgres = "^4.0"
sqlalchemy = ">=1.4.46,<1.5.0"
pymysql = "^1.0.2"
sqlalchemy-utils = "^0.39.0"
psycopg2-binary = "^2.9.5"
flask-sqlalchemy = "2.5.1"
python-dotenv = ">=0.21.0,<0.22.0"
flask = ">=2.2.2,<2.3.0"
flask-script = "2.0.6"
requests = "^2.28.2"
pytest-cov = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
